python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.36",
    "botocore>=1.42.36",
    "dotenv>=0.9.9",
    "fastapi>=0.131.0",
    "kaggle>=1.8.3",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "streamlit>=1.54.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.41.0",
]
